+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/torState/TorRelayEvaluation.kt
@@ -20,30 +20,5 @@
  */
 package com.vitorpamplona.amethyst.model.torState
 
-import com.vitorpamplona.amethyst.commons.tor.TorType
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.isLocalHost
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.isOnion
-
-class TorRelayEvaluation(
-    val torSettings: TorRelaySettings,
-    val trustedRelayList: Set<NormalizedRelayUrl>,
-    val dmRelayList: Set<NormalizedRelayUrl>,
-) {
-    fun useTor(relay: NormalizedRelayUrl): Boolean =
-        if (torSettings.torType == TorType.OFF) {
-            false
-        } else {
-            if (relay.isLocalHost()) {
-                false
-            } else if (relay.isOnion()) {
-                torSettings.onionRelaysViaTor
-            } else if (relay in dmRelayList) {
-                torSettings.dmRelaysViaTor
-            } else if (relay in trustedRelayList) {
-                torSettings.trustedRelaysViaTor
-            } else {
-                torSettings.newRelaysViaTor
-            }
-        }
-}
+// Canonical type now lives in commons
+typealias TorRelayEvaluation = com.vitorpamplona.amethyst.commons.tor.TorRelayEvaluation
